# Conflicts:
#	app/src/main/java/rconnect/retvens/technologies/dashboard/configuration/roomsAndRates/amenity/CreateAmenityDialog.kt

+++ app/src/main/java/rconnect/retvens/technologies/dashboard/channelManager/AddReservationFragment/RoomDetailPrices.kt
@@ -3,10 +3,10 @@ package rconnect.retvens.technologies.dashboard.channelManager.AddReservationFra
 data class RoomDetailPrices(
     var baseChildCount:Int = 0,
     var baseAdultCount:Int = 0,
-    var pricePerExtraAdult:Int = 0,
-    var pricePerExtraChild:Int = 0,
-    var totalCharges:Int = 0,
-    var totalPriceForExtraChild:Int = 0,
-    var  totalPriceForExtraAdult:Int = 0,
+    var pricePerExtraAdult:Double = 0.0,
+    var pricePerExtraChild:Double = 0.0,
+    var totalCharges:Double = 0.0,
+    var totalPriceForExtraChild:Double = 0.0,
+    var  totalPriceForExtraAdult:Double = 0.0,
 ) {
 }
